+++ doc/whats_new.rst
@@ -127,6 +127,11 @@ Enhancements
      A ``TypeError`` will be raised for any other kwargs. :issue:`8028`
      by :user:`Alexander Booth <alexandercbooth>`.
 
+   - Used Embarrassingly Parallel helper in the :func:`fit` in
+     :class:`gaussian_process.GaussianProcessRegressor` and tested the same for
+     different ``n_jobs`` parameter values. Also added Timing Benchmark for
+     differnet n_jobs values. :issue:`7987` by :user:`Aman Pratik <amanp10>`
+
    - Added type checking to the ``accept_sparse`` parameter in
      :mod:`sklearn.utils.validation` methods. This parameter now accepts only
      boolean, string, or list/tuple of strings. ``accept_sparse=None`` is deprecated